requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
build==0.8.0
    # via pip-tools
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via pip-tools
idna==3.3
    # via requests
numpy==1.23.1
    # via torchvision
packaging==21.3
    # via build
pep517==0.13.0
    # via build
pillow==9.2.0
    # via
    #   -r requirements.in
    #   torchvision
pip-tools==6.8.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
requests==2.28.1
    # via torchvision
tomli==2.0.1
    # via
    #   build
    #   pep517
torch==1.12.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.13.0
    # via -r requirements.in
tqdm==4.64.0
    # via -r requirements.in
typing-extensions==4.3.0
    # via
    #   torch
    #   torchvision
urllib3==1.26.11
    # via requests
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
